requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.6
django-appconf==1.1.0
django-axes==5.0
django-environ==0.12.0
django-extensions==3.2.1
django-ipware==7.0.1
django-recaptcha==3.0.0
djangorestframework==3.16.1
gunicorn==20.1.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ipware==3.0.0
redis==6.4.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
django-auth-ldap==5.0.0
python-ldap==3.4.4
django-widget-tweaks==1.5.0
Pillow==10.4.0
